+++ fatman/src/main/java/UMCFatMan/fatman/domain/weekRank/DTO/WeekRankResponseDto.java
@@ -1,6 +1,6 @@
 package UMCFatMan.fatman.domain.weekRank.DTO;
 
-import UMCFatMan.fatman.domain.users.Users;
+import UMCFatMan.fatman.domain.users.entity.Users;
 import UMCFatMan.fatman.domain.weekRank.WeekRank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
